requirements_client.txt
grpcio~=1.34.0
protobuf~=3.14.0
pandas>=1.0.0